Pre-test tweaking of similar subsampling. One file for one subsampling purpose.
