+++ og_marl/vault_utils/subsample_similar.py
@@ -1,26 +1,13 @@
 import jax
 import jax.numpy as jnp
 import numpy as np
-import flashbax as fbx
 from flashbax.vault import Vault
-from scipy.stats import norm
-import matplotlib.pyplot as plt
-import copy
-import flashbax
-from flashbax.buffers.trajectory_buffer import TrajectoryBufferState
-
-
-# just gathering good, medium and poor vaults in a list of unread vaults
-def get_all_vaults(rel_dir,vault_name,vault_uids=["Good","Medium","Poor"]):
-    vlts = []
-    for quality in vault_uids:
-        vlts.append(Vault(rel_dir=rel_dir, vault_name=vault_name, vault_uid=quality))
-    return vlts
+from og_marl.vault_utils.subsample_smaller import stitch_vault_from_sampled_episodes_
+from og_marl.vault_utils.download_vault import check_directory_exists_and_not_empty
+import pickle
 
 # cumulative summing per-episode
-def get_episode_returns_and_term_idxes(vlt):
-    all_data = vlt.read()
-    offline_data = all_data.experience
+def get_episode_returns_and_term_idxes(offline_data):
 
     rewards = offline_data['rewards'][0, :, 0]
     terminal_flag = offline_data['terminals'][0, :, ...].all(axis=-1)
@@ -41,116 +28,21 @@ def get_episode_returns_and_term_idxes(vlt):
     # shift back as we did for the loop
     return cumsums[term_idxes-1], term_idxes
 
+
 # first store indices of episodes, then sort by episode return.
 # outputs return, start, end and vault index in vault list
-def sort_concat(returns,eps_ends, ids):
+def sort_concat(returns,eps_ends):
 
     episode_start_idxes = eps_ends[:-1]+1
     episode_start_idxes = jnp.insert(episode_start_idxes,0,0).reshape(-1,1)
     sorting_idxes = jnp.lexsort(jnp.array([returns[:,0]]), axis=-1)
     # print(sorting_idxes)
 
-    return_start_end = jnp.concatenate([returns,episode_start_idxes.reshape(-1,1),eps_ends,ids],axis=-1)
+    return_start_end = jnp.concatenate([returns,episode_start_idxes.reshape(-1,1),eps_ends],axis=-1)
 
     # return, start, end sorted by return value ascending
     sorted_return_start_end = return_start_end[sorting_idxes]
     return sorted_return_start_end
-
-# given bin edges and a sorted array of values, get the bin number per value
-def get_bin_numbers(sorted_values,bin_edges):
-    bin_numbers = np.zeros_like(sorted_values)
-    
-    def get_bin_number(bin_num,value):
-        is_overflowing = value>bin_edges[bin_num]
-
-        if is_overflowing:
-            bin_num += 1
-            is_doubly_overflowing = value>bin_edges[bin_num]
-            if is_doubly_overflowing:
-                bin_num = get_bin_number(bin_num,value)
-            
-        return bin_num
-    
-    bin_bookmark = 0
-
-    for i,val in enumerate(sorted_values):
-        bin_bookmark = get_bin_number(bin_bookmark,val)
-        bin_numbers[i] = bin_bookmark
-
-    return bin_numbers
-
-# sample from pdf according to heights
-# BIG NOTE: CHECK THE DISPARITY, OTHERWISE YOUR DISTRIBUTION WILL BE TOO MUCH
-def episode_idxes_sampled_from_pdf(pdf,bar_heights):
-    num_to_sample = np.round(pdf).astype(int)
-    sample_range_edges = np.concatenate([[0],np.cumsum(bar_heights)])
-
-    assert num_to_sample.shape==bar_heights.shape
-
-    target_sample_idxes = []
-    for i,n_sample in enumerate(num_to_sample):
-            sample_base = np.arange(sample_range_edges[i],sample_range_edges[i+1])
-            print(sample_base)
-            if n_sample<=0:
-                 pass
-            # if we sample more than all in the bar
-            else:
-                if n_sample>=bar_heights[i]:
-                    sample_rest = np.random.choice(sample_base,n_sample-bar_heights[i],replace=True)
-                    sample = np.concatenate([sample_base,sample_rest])
-                else:
-                    sample = np.random.choice(sample_base,n_sample,replace=True) #make false for no replace
-                target_sample_idxes = target_sample_idxes+list(np.sort(sample))
-    return target_sample_idxes
-
-def preprocess(rel_dir,vault_name):
-    vlts = get_all_vaults(rel_dir, vault_name)
-
-    # get returns, term idxes for each episode per vault
-    returns_list = []
-    episode_end_list = []
-    vault_ids = []
-    for j,vault in enumerate(vlts):
-        print(j)
-        returns, episode_end_idxes = get_episode_returns_and_term_idxes(vault)
-        returns_list.append(returns)
-        episode_end_list.append(episode_end_idxes)
-        vault_ids.append(jnp.zeros_like(returns)+j)
-
-    # make np compatible
-    all_returns = jnp.concatenate(returns_list)
-    all_episode_end_idxes = jnp.concatenate(episode_end_list)
-    all_vault_ids = jnp.concatenate(vault_ids)
-
-    # concatenate then sort all results
-    all_sorted_return_start_end = sort_concat(all_returns,all_episode_end_idxes, all_vault_ids)
-    return vlts, all_sorted_return_start_end
-
-def preprocess_vault(rel_dir,vault_name, vault_uid):
-    vault = Vault(rel_dir=rel_dir, vault_name=vault_name,vault_uid=vault_uid)
-
-    returns, episode_end_idxes = get_episode_returns_and_term_idxes(vault)
-    
-    all_sorted_return_start_end = sort_concat(returns,episode_end_idxes, jnp.zeros_like(returns))
-    return vault, all_sorted_return_start_end
-
-def bin_processed_data(all_sorted_return_start_end, n_bins=500):
-    # get bin edges, including final endpoint
-    bin_edges = jnp.linspace(start=min(min(all_sorted_return_start_end[:,0]),0), stop = max(all_sorted_return_start_end[:,0]), num=n_bins,endpoint=True)
-    print(all_sorted_return_start_end.shape[0])
-
-    # get bin numbers
-    bin_numbers = get_bin_numbers(all_sorted_return_start_end[:,0], bin_edges)
-    print(bin_numbers.shape[0])
-
-    bar_labels, bar_heights= np.unique(bin_numbers,return_counts=True)
-
-    padded_heights = np.zeros(n_bins)
-    for bar_l, bar_h in zip(bar_labels,bar_heights):
-        padded_heights[int(bar_l)] = bar_h
-
-    return bar_labels, bar_heights, padded_heights, bin_edges, bin_numbers
-
 
 def get_idxes_of_similar_subsets(base_returns, comp_returns, tol=0.1):
     base_selected_idxes = []
@@ -170,3 +62,65 @@ def get_idxes_of_similar_subsets(base_returns, comp_returns, tol=0.1):
             comp_idx+= 1
 
     return base_selected_idxes, comp_selected_idxes
+
+
+def subsample_similar(first_vault_info, second_vault_info, new_rel_dir, new_vault_name):
+
+    # check that a subsampled vault by the same name does not already exist
+    if check_directory_exists_and_not_empty(f"./{new_rel_dir}/{new_vault_name}"):
+        print(f"Vault '{new_rel_dir}/{new_vault_name.strip('.vlt')}' already exists. To subsample from scratch, please remove the current subsampled vault from its directory.")
+        return
+
+    # store the return start end info
+    return_start_end_list = []
+    experience_list = []
+    for vault_info in [first_vault_info, second_vault_info]:
+
+        # unpack info
+        rel_dir = vault_info['rel_dir']
+        vault_name = vault_info['vault_name']
+        vault_uid = vault_info['uid']
+
+        # check that the vault to be subsampled exists
+        if not check_directory_exists_and_not_empty(f"./{rel_dir}/{vault_name}"):
+            print(f"Vault './{rel_dir}/{vault_name}' does not exist and cannot be subsampled.")
+            return
+        
+        # get the vault and unpack its experience
+        vlt = Vault(rel_dir=rel_dir, vault_name=vault_name,vault_uid=vault_uid)
+
+        all_data = vlt.read()
+        offline_data = all_data.experience
+        del all_data
+
+        # basic information about the vault
+        returns, episode_end_idxes = get_episode_returns_and_term_idxes(offline_data)
+    
+        # sort the episodes by return and store them
+        return_start_end_list.append(sort_concat(returns,episode_end_idxes))
+        experience_list.append(offline_data)
+
+
+    base_ret = return_start_end_list[0][:,0]
+    comp_ret = return_start_end_list[1][:,0]
+
+    b,c = get_idxes_of_similar_subsets(base_ret,comp_ret,tol=0.01)
+
+    vlt1_samples = return_start_end_list[0][b,:]
+    vlt2_samples = return_start_end_list[1][c,:]
+
+    for experience,sample_idxes,uid in zip(experience_list,[vlt1_samples,vlt2_samples],[first_vault_info['uid'], second_vault_info['uid']]):
+
+        timesteps_written = stitch_vault_from_sampled_episodes_(
+                                                experience=experience,
+                                                len_start_end_sample=sample_idxes,
+                                                dest_vault_name=new_vault_name,
+                                                vault_uid=uid,
+                                                rel_dir=new_rel_dir,
+                                                )
+        
+        # save the number of timesteps actually written
+        with open(f"{new_rel_dir}/{new_vault_name}/{uid}/timesteps.pickle","wb") as f:
+            pickle.dump(timesteps_written,f)
+
+    return
